+++ dessia_common/models/workflows/forms_workflow.py
@@ -9,17 +9,15 @@ instanciate_generator = InstantiateModel(model_class=Generator, name='Instantiat
 
 generate_method = MethodType(class_=Generator, name='generate')
 generator_generate = ModelMethod(method_type=generate_method, name='Generator Generate')
-generate_attribute = AttributeType(class_=Generator, name='models')
-attribute_selection = ModelAttribute(attribute_type=generate_attribute, name='Attribute Selection')
+attribute_selection = ModelAttribute(attribute_name='models', name='Attribute Selection')
 
 # Sub-Workflow of model optimization
 instanciate_optimizer = InstantiateModel(model_class=Optimizer, name='Instantiate Optimizer')
 
-
 generate_method = MethodType(class_=Optimizer, name='optimize')
 optimization = ModelMethod(method_type=generate_method, name='Optimization')
-model_to_optimize_attribute = AttributeType(class_=Optimizer, name='model_to_optimize')
-model_fetcher = ModelAttribute(attribute_type=model_to_optimize_attribute, name='Model Fetcher')
+
+model_fetcher = ModelAttribute(attribute_name='model_to_optimize', name='Model Fetcher')
 
 pipe1_opt = Pipe(input_variable=instanciate_optimizer.outputs[0], output_variable=optimization.inputs[0])
 pipe2_opt = Pipe(input_variable=optimization.outputs[1], output_variable=model_fetcher.inputs[0])